fix(compiler): regex/super/dynamic-import in async & generator bodies (#1105)
The four state-machine emitters derive from StatementEmitterBase, not
ILEmitter, so a regex literal compiled to `null` in any async/generator
body (the base EmitRegexLiteral pushed Ldnull), and the generator emitters
additionally overrode EmitSuper/EmitDynamicImport to push `null`. Result:
`/re/`, `super.x`, and `import()` inside an `async function`/`function*`
silently evaluated to null at runtime — wrong result, no error.

- Port the hoist-aware $RegExp emission from ILEmitter into the shared
  ExpressionEmitterBase (the per-site hoist fields already hang off the
  shared EmittedRuntime, so it needs no sync-only state). Regex now
  compiles identically in functions, arrows, and all four state machines.
  Drop the now-redundant ILEmitter override.
- Remove the silent-null EmitSuper overrides in GeneratorMoveNextEmitter
  and AsyncGeneratorMoveNextEmitter; the base default (hoisted `this` ->
  GetSuperMethod) works in the state machine. `super.x` now resolves in
  generators and async generators, matching async functions.
- Remove the silent-null EmitDynamicImport override in
  GeneratorMoveNextEmitter; the base default routes through the module
  registry, behaving identically to async functions (the remaining
  "module not pre-compiled" bundling limitation is orthogonal and not
  state-machine-specific).
- Update EmitterSyncTests allowlist (the removed overrides) and add 8
  differential-parity snippets pinning regex/super across the four bodies.

Suite 14519/0, EmitterSyncTests + DifferentialParityTests green; verified
interp==compiled for regex/super in every state-machine body.
