+++ Compilation/AsyncGeneratorMoveNextEmitter.Expressions.cs
@@ -767,14 +767,8 @@ public partial class AsyncGeneratorMoveNextEmitter
 
     #endregion
 
-    #region Missing Abstract Implementations
-
-    protected override void EmitSuper(Expr.Super s)
-    {
-        // Super not supported in async generators - push null
-        _il.Emit(OpCodes.Ldnull);
-        SetStackUnknown();
-    }
-
-    #endregion
+    // EmitSuper is inherited from ExpressionEmitterBase (#1105): the base loads the
+    // hoisted `this` and resolves through GetSuperMethod, which works in the async
+    // generator state machine. The old override here pushed `null`, silently
+    // miscompiling `super.x` inside an async generator body.
 }
